requirements.txt
argon2-cffi
asgiref==3.3.1
async-generator
attrs
backcall
bleach
certifi==2020.12.5
cffi
chardet==4.0.0
click==7.1.2
colorama
cryptography==3.4.6
cycler==0.10.0
decorator
defusedxml
Django==3.1.7
docx2txt==0.8
entrypoints==0.3
Flask==1.1.2
idna==2.10
importlib-metadata
ipykernel==5.5.0
ipython
ipython-genutils
itsdangerous==1.1.0
jedi==0.18.0
Jinja2
joblib==1.0.1
jsonschema
jupyter-client==6.1.11
jupyter-core
jupyterlab-pygments
kiwisolver==1.3.1
MarkupSafe
matplotlib==3.3.4
mistune
nbclient
nbconvert
nbformat
nest-asyncio
nltk==3.5
notebook
numpy==1.20.1
packaging
pandas==1.2.3
pandocfilters
parso==0.8.1
pdfminer.six==20201018
pickleshare
Pillow==8.1.1
prometheus-client
prompt-toolkit==3.0.16
pycparser
Pygments
pyparsing
pyrsistent
python-dateutil
pytz==2021.1
pywin32==300
pywinpty
pyzmq==22.0.3
regex==2020.11.13
requests==2.25.1
Send2Trash
six
sortedcontainers==2.3.0
sqlparse==0.4.1
terminado==0.9.2
testpath
tornado
tqdm==4.58.0
traitlets
urllib3==1.26.3
wcwidth
webencodings==0.5.1
Werkzeug==1.0.1
wincertstore==0.2
wordcloud==1.8.1
zipp
